% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{aggregate_cull_data}
\alias{aggregate_cull_data}
\title{Aggregate Cull Data}
\usage{
aggregate_cull_data(cull_data)
}
\arguments{
\item{cull_data}{Prepared cull data.}
}
\value{
Summary dataset by Year and ReefLabel.
}
\description{
Aggregate Cull Data
}
