% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{best_pair}
\alias{best_pair}
\title{Best Pair Threshold}
\usage{
best_pair(df, cpue_thr, perc_grid, conc_grid_t, rule = "and")
}
\arguments{
\item{df}{Data}

\item{cpue_thr}{Target CPUE}

\item{perc_grid}{Grid}

\item{conc_grid_t}{Grid}

\item{rule}{logic}
}
\value{
list
}
\description{
Best Pair Threshold
}
