% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_mesh.R
\name{build_inla_spde_mesh}
\alias{build_inla_spde_mesh}
\title{Build a projected SPDE mesh for COTS CPUE modelling}
\usage{
build_inla_spde_mesh(
  locations,
  cutoff_m = 5000,
  max_edge_m = c(20000, 80000),
  offset_m = c(20000, 1e+05)
)
}
\arguments{
\item{locations}{An \code{sf} point object in a projected metre-based CRS.}

\item{cutoff_m}{Minimum separation between retained mesh input locations.}

\item{max_edge_m}{Inner and outer maximum triangle edge lengths.}

\item{offset_m}{Inner and outer mesh extension distances.}
}
\value{
An INLA two-dimensional mesh.
}
\description{
Build a projected SPDE mesh for COTS CPUE modelling
}
