% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{build_popup}
\alias{build_popup}
\title{Build Leaflet Popup Html}
\usage{
build_popup(
  reef,
  year,
  edna_val,
  scar_val,
  cpue_val,
  truth,
  conf_class,
  method
)
}
\arguments{
\item{reef}{Reef name}

\item{year}{Year}

\item{edna_val}{eDNA metric}

\item{scar_val}{Scar metric}

\item{cpue_val}{CPUE metric}

\item{truth}{Logical true outcome}

\item{conf_class}{Confusion classification string}

\item{method}{Description of prediction method}
}
\value{
String HTML representation
}
\description{
Build Leaflet Popup Html
}
