% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{combine_reef_data}
\alias{combine_reef_data}
\title{Combine All Reef Data}
\usage{
combine_reef_data(cull_summary, first_day_summary, manta_summary, edna_summary)
}
\arguments{
\item{cull_summary}{Checked and evaluated summarizations for cull.
For temporally correct analysis, cull data should be pre-filtered
via \code{\link[=filter_cull_post_edna]{filter_cull_post_edna()}} before aggregation.}

\item{first_day_summary}{Scout dive metrics.}

\item{manta_summary}{Evaluated Manta metrics.}

\item{edna_summary}{Summary of eDNA traces.}
}
\value{
Joined wide dataset combining all field campaigns.
}
\description{
Combine All Reef Data
}
