% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{evaluate_unified_spatial}
\alias{evaluate_unified_spatial}
\title{Evaluate Unified Spatial}
\usage{
evaluate_unified_spatial(df, cp, perc_grid)
}
\arguments{
\item{df}{Spatial data frame}

\item{cp}{CPUE target}

\item{perc_grid}{Grid array}
}
\value{
list
}
\description{
Evaluate Unified Spatial
}
