% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_models.R
\name{extract_inla_cpue_predictions}
\alias{extract_inla_cpue_predictions}
\title{Extract fitted or scenario CPUE posterior summaries}
\usage{
extract_inla_cpue_predictions(model, tag = c("pred", "est"))
}
\arguments{
\item{model}{A fitted \code{reefdna_inla_benchmark} object.}

\item{tag}{Either \code{est} for fitted observations or \code{pred} for scenarios.}
}
\value{
An \code{sf} object containing posterior response summaries and CPUE.
}
\description{
Extract fitted or scenario CPUE posterior summaries
}
