% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{get_scout_metrics}
\alias{get_scout_metrics}
\title{Get Scout Metrics}
\usage{
get_scout_metrics(cull_data)
}
\arguments{
\item{cull_data}{Prepared cull data.}
}
\value{
Early scout metrics for the first day on a reef.
}
\description{
Get Scout Metrics
}
