% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{lm_label}
\alias{lm_label}
\title{Format Linear Model Label for Plotting}
\usage{
lm_label(model, digits_r2 = 2, digits_p = 3, digits_rmse = 2)
}
\arguments{
\item{model}{A fitted linear model object.}

\item{digits_r2}{Digits for R squared}

\item{digits_p}{Digits for p-value}

\item{digits_rmse}{Digits for RMSE}
}
\value{
Formatted string label
}
\description{
Format Linear Model Label for Plotting
}
