% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{make_confusion}
\alias{make_confusion}
\title{Create Confusion Matrix Object}
\usage{
make_confusion(dat, perc_thresh, cpue_thresh)
}
\arguments{
\item{dat}{Data containing \code{perc_pos} and \code{cpue}.}

\item{perc_thresh}{Percentage positive threshold.}

\item{cpue_thresh}{CPUE target threshold.}
}
\value{
A list containing the confusion matrix data frame (\code{cm}) and computed metrics.
}
\description{
Create Confusion Matrix Object
}
