% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{make_confusion_2d}
\alias{make_confusion_2d}
\title{Make 2D Confusion Matrix}
\usage{
make_confusion_2d(dat, perc_star, conc_t_star, cpue_thr, rule = "and")
}
\arguments{
\item{dat}{Dataframe}

\item{perc_star}{Chosen perc pos threshold}

\item{conc_t_star}{Chosen conc threshold}

\item{cpue_thr}{CPUE threshold}

\item{rule}{"and" or "or"}
}
\value{
List with cm
}
\description{
Make 2D Confusion Matrix
}
