% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{make_popup}
\alias{make_popup}
\title{Make Leaflet Popup Html for Droppoints}
\usage{
make_popup(reef, date_edna, perc_pos, conc_mean, n_samples)
}
\arguments{
\item{reef}{Reef Name}

\item{date_edna}{Date}

\item{perc_pos}{Percent positive}

\item{conc_mean}{Mean concentration}

\item{n_samples}{Sample count}
}
\value{
String HTML
}
\description{
Make Leaflet Popup Html for Droppoints
}
