% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics_for}
\alias{metrics_for}
\title{Evaluate Performance Metrics}
\usage{
metrics_for(p_thr, c_thr, x)
}
\arguments{
\item{p_thr}{Numeric percentage positive threshold}

\item{c_thr}{Numeric CPUE threshold}

\item{x}{Dataframe with \code{perc_pos} and \code{cpue} columns}
}
\value{
Named numeric vector containing F1, Accuracy, Precision, and Recall.
}
\description{
Evaluate Performance Metrics
}
