% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metrics.R
\name{metrics_for_both}
\alias{metrics_for_both}
\title{Metrics for Both}
\usage{
metrics_for_both(p_thr, c_thr, cpue_thr, rule = c("or", "and"), x)
}
\arguments{
\item{p_thr}{Numeric percentage positive threshold}

\item{c_thr}{Numeric Concentration threshold}

\item{cpue_thr}{CPUE target}

\item{rule}{Boolean logic rule: "or" or "and"}

\item{x}{Dataframe with \code{perc_pos}, \code{conc_t} and \code{cpue}}
}
\value{
Numeric performance metrics
}
\description{
Metrics for Both
}
