% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_data.R
\name{normalize_cull_site_id}
\alias{normalize_cull_site_id}
\title{Normalize a cull-site identifier}
\usage{
normalize_cull_site_id(x)
}
\arguments{
\item{x}{Character vector containing cull-site names.}
}
\value{
Upper-case alphanumeric identifiers.
}
\description{
Normalize a cull-site identifier
}
