% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz.R
\name{plot_confusion}
\alias{plot_confusion}
\title{Plot Confusion Matrix}
\usage{
plot_confusion(cm_obj, title = "Confusion matrix")
}
\arguments{
\item{cm_obj}{Confusion matrix list object}

\item{title}{Plot title}
}
\value{
ggplot object
}
\description{
Plot Confusion Matrix
}
