% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{prepare_cull_data}
\alias{prepare_cull_data}
\title{Prepare Cull Data}
\usage{
prepare_cull_data(cull_file)
}
\arguments{
\item{cull_file}{Path to the cull data excel file.}
}
\value{
Cleaned cull dataset with CPUE and dive class labels.
}
\description{
Prepare Cull Data
}
