% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{prepare_manta_data}
\alias{prepare_manta_data}
\title{Prepare Manta Tow Data}
\usage{
prepare_manta_data(manta_file)
}
\arguments{
\item{manta_file}{Path to manta tow csv file.}
}
\value{
Cleaned manta tow dataset.
}
\description{
Prepare Manta Tow Data
}
