% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{process_edna_events}
\alias{process_edna_events}
\title{Process eDNA Survey Events}
\usage{
process_edna_events(edna_data, gap_days = 7)
}
\arguments{
\item{edna_data}{Raw eDNA data.}

\item{gap_days}{Maximum days between samples to be considered the same event.}
}
\value{
Aggregated event-level dataset.
}
\description{
Process eDNA Survey Events
}
