% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{run_cv_2d}
\alias{run_cv_2d}
\title{Run 2D Cross Validation}
\usage{
run_cv_2d(dat, cpue_thr, perc_grid, conc_grid_t, rule = "and", k = 5, R = 100)
}
\arguments{
\item{dat}{Dataset}

\item{cpue_thr}{CPUE target}

\item{perc_grid}{Grid 1}

\item{conc_grid_t}{Grid 2}

\item{rule}{Boolean rule}

\item{k}{Folds}

\item{R}{repeats}
}
\value{
list
}
\description{
Run 2D Cross Validation
}
