% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{run_cv_for_cpue}
\alias{run_cv_for_cpue}
\title{Run CV for CPUE threshold}
\usage{
run_cv_for_cpue(dat_evt, cpue_thr, perc_grid, k = 5, R = 100)
}
\arguments{
\item{dat_evt}{Event dataset}

\item{cpue_thr}{CPUE Threshold}

\item{perc_grid}{Grid}

\item{k}{Folds}

\item{R}{Repeats}
}
\value{
List of performance
}
\description{
Run CV for CPUE threshold
}
