% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{run_cv_for_horizon}
\alias{run_cv_for_horizon}
\title{Run CV for Horizon}
\usage{
run_cv_for_horizon(dat_full, hor_val, cpue_thr, perc_grid, k = 5, R = 100)
}
\arguments{
\item{dat_full}{Full dataset}

\item{hor_val}{Horizon}

\item{cpue_thr}{Target cpue}

\item{perc_grid}{Grid}

\item{k}{folds}

\item{R}{repeats}
}
\value{
List of performance
}
\description{
Run CV for Horizon
}
