% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inla_mesh.R
\name{summarise_inla_mesh}
\alias{summarise_inla_mesh}
\title{Summarize an INLA SPDE mesh}
\usage{
summarise_inla_mesh(mesh)
}
\arguments{
\item{mesh}{An INLA mesh.}
}
\value{
A one-row tibble of mesh dimensions and edge-length summaries.
}
\description{
Summarize an INLA SPDE mesh
}
