% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{summarize_edna_data}
\alias{summarize_edna_data}
\title{Summarize eDNA Data}
\usage{
summarize_edna_data(edna_file, sheet = "eDNA_data_ALL")
}
\arguments{
\item{edna_file}{Path to eDNA excel file}

\item{sheet}{Sheet name, defaults to "eDNA_data_ALL"}
}
\value{
Summarised eDNA data per reef/year, including earliest eDNA_Date.
}
\description{
Summarize eDNA Data
}
