% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{summarize_manta_data}
\alias{summarize_manta_data}
\title{Summarize Manta Tow Data}
\usage{
summarize_manta_data(manta_data)
}
\arguments{
\item{manta_data}{Cleaned manta tow data}
}
\value{
Summary by Reef and Year.
}
\description{
Summarize Manta Tow Data
}
