% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{fit_horizon_glmm}
\alias{fit_horizon_glmm}
\title{Fit Horizon GLMM}
\usage{
fit_horizon_glmm(hor_val, dat_glmm)
}
\arguments{
\item{hor_val}{Horizon Value ("3 Months", etc)}

\item{dat_glmm}{Dataset containing modelling inputs.}
}
\value{
Predictions tibble
}
\description{
Fit Horizon GLMM
}
