% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/modelling.R
\name{get_prior_cohort}
\alias{get_prior_cohort}
\title{Get Prior Cohort}
\usage{
get_prior_cohort(culls, ednas, win_days, label)
}
\arguments{
\item{culls}{Cull dataset}

\item{ednas}{eDNA dataset}

\item{win_days}{Window}

\item{label}{Label string}
}
\value{
Data frame
}
\description{
Get Prior Cohort
}
